requirements.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.3.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.3
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.104.1
greenlet==3.2.3
h11==0.16.0
idna==3.10
loguru==0.7.2
pip==21.2.3
pyasn1==0.6.1
pycparser==2.22
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.10.1
PyMySQL==1.1.0
python-dotenv==1.0.0
rsa==4.9.1
setuptools==57.4.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.14.0
typing-inspection==0.4.1
uvicorn==0.24.0
